feat(abtesting/optimizelyXExperiment): better organize component file structure
