+++ components/abtesting/optimizelyXExperiment/src/experiment-provider-only/index.js
@@ -1,0 +1,19 @@
+import React from 'react'
+import PropTypes from 'prop-types'
+
+import ExperimentContext from '../experiment-context'
+
+function ExperimentProviderOnly({children, experimentData}) {
+  return (
+    <ExperimentContext.Provider value={experimentData}>
+      {children}
+    </ExperimentContext.Provider>
+  )
+}
+
+ExperimentProviderOnly.propTypes = {
+  children: PropTypes.any.isRequired,
+  experimentData: PropTypes.object
+}
+
+export default ExperimentProviderOnly
